requirements.txt
openai-whisper
pyinstaller
ffmpeg-python
